% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_input.R
\name{choose_file_input_dialog}
\alias{choose_file_input_dialog}
\title{Input file modalDialog chooser}
\usage{
choose_file_input_dialog(initial = FALSE)
}
\arguments{
\item{initial}{logical indicating wheter it is the initial fileInput or an update}
}
\value{
shiny::modalDialog object
}
\description{
Helper to get a consistent dialog modal for the fileInput
}
\author{
Emanuel Sommer
}
