% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_largest_streak}
\alias{get_largest_streak}
\title{Get Largest streak from POSIXct vector}
\usage{
get_largest_streak(time_vec)
}
\arguments{
\item{time_vec}{POSIXct vector (sorted, length > 1 and no missing values)}
}
\value{
numeric value the largest streak
}
\description{
i.e. most consecutive days in a row (Implementation is quick and dirty
probably not the most efficient one, one pass through unique days)
}
\author{
Emanuel Sommer
}
