% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_conversation_starter}
\alias{plot_conversation_starter}
\title{Barplot displaying the relative frequencies of the person who starts
the conversations}
\usage{
plot_conversation_starter(data)
}
\arguments{
\item{data}{tibble provided by the \code{prep_data()} function}
}
\value{
ggplot2 object
}
\description{
Barplot displaying the relative frequencies of the person who starts
the conversations
}
\author{
Eleftheria Papavasileiou
}
