% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_day_activity}
\alias{plot_day_activity}
\title{Barplot displaying the chat activity for a specific day of the week}
\usage{
plot_day_activity(
  data,
  days = c("Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday",
    "Sunday")
)
}
\arguments{
\item{data}{tibble provided by the \code{prep_data()} function}

\item{days}{character vector of weekdays ("Monday", "Tuesday", "Wednesday",
"Thursday", "Friday", "Saturday", "Sunday")}
}
\value{
ggplot2 object
}
\description{
Barplot displaying the chat activity for a specific day of the week
}
\author{
Emanuel Sommer
}
