% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_emoji_words_per_mess_box}
\alias{plot_emoji_words_per_mess_box}
\title{Boxplot of emojis or words per message (distinct users)}
\usage{
plot_emoji_words_per_mess_box(data, emo = TRUE)
}
\arguments{
\item{data}{tibble provided by the \code{prep_data()} function}

\item{emo}{logical whether to plot emojis or words per message}
}
\value{
ggplot2 object
}
\description{
Boxplot of emojis or words per message (distinct users)
}
\author{
Emanuel Sommer
}
