% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_emoji_words_per_mess_dens}
\alias{plot_emoji_words_per_mess_dens}
\title{Density  plot of emojis or words per message (distinct users by fill/color)}
\usage{
plot_emoji_words_per_mess_dens(data, emo = TRUE, bw = NULL)
}
\arguments{
\item{data}{tibble provided by the \code{prep_data()} function}

\item{emo}{logical whether to plot emojis or words per message}

\item{bw}{bandwidth of the gaussian kernels (default: \code{stats::bw.nrd0()})}
}
\value{
ggplot2 object
}
\description{
Density  plot of emojis or words per message (distinct users by fill/color)
}
\author{
Eleftheria Papavasileiou
}
