% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_last_man_standing}
\alias{plot_last_man_standing}
\title{Barplot displaying the relative frequencies of the person is the last person
in the chat}
\usage{
plot_last_man_standing(data)
}
\arguments{
\item{data}{tibble provided by the \code{prep_data()} function}
}
\value{
ggplot2 object
}
\description{
frequency of last message before a conversation break of at least 4 days
}
\author{
Eleftheria Papavasileiou
}
