% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_top10_emojis}
\alias{plot_top10_emojis}
\title{Barchart to display the frequency of the top 10 used emojis of a given user}
\usage{
plot_top10_emojis(data, authors)
}
\arguments{
\item{data}{tibble provided by the \code{prep_data()} function}

\item{authors}{character or factor vector conatining valid authors from the data$author column}
}
\value{
ggplot2 object
}
\description{
Barchart to display the frequency of the top 10 used emojis of a given user
}
\author{
Emanuel Sommer
}
