% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_total_words}
\alias{plot_total_words}
\title{Barplot displaying the relative frequencies of the users}
\usage{
plot_total_words(data)
}
\arguments{
\item{data}{tibble provided by the \code{prep_data()} function}
}
\value{
ggplot2 object
}
\description{
Barplot displaying the relative frequencies of the users
}
\author{
Emanuel Sommer
}
