% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_ts_mess_per_day}
\alias{plot_ts_mess_per_day}
\title{PLot time against messages sent per day (scatterplot + smoothing line for trends)}
\usage{
plot_ts_mess_per_day(data)
}
\arguments{
\item{data}{tibble provided by the \code{prep_data()} function}
}
\value{
plotly object
}
\description{
PLot time against messages sent per day (scatterplot + smoothing line for trends)
}
\author{
Emanuel Sommer
}
