% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_weekday_activity}
\alias{plot_weekday_activity}
\title{Barplot portraying the chat activity by weekday}
\usage{
plot_weekday_activity(data)
}
\arguments{
\item{data}{tibble provided by the \code{prep_data()} function}
}
\value{
ggplot2 object
}
\description{
Barplot portraying the chat activity by weekday
}
\author{
Emanuel Sommer
}
