% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_prep.R
\name{prep_data}
\alias{prep_data}
\title{prep the data after loading}
\usage{
prep_data(rwa_data)
}
\arguments{
\item{rwa_data}{output tibble of \code{rwhatsapp::rwa_read()}}
}
\value{
augmented tibble (remove: source, add: n_words, n_emojis, wday)
}
\description{
removes unnecessary and adds useful columns to the input data
}
\details{
Remove:
\itemize{
\item source column
}

Add:
\itemize{
\item \code{n_words} word count per message
\item \code{n_emojis} emoji count per message
\item \code{wday} weekday (integer, 1 corresponds to Sunday)
}
}
\author{
Emanuel Sommer / Eleftheria Papavasileiou
}
