% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{test_chat_data}
\alias{test_chat_data}
\title{A tibble containing test data as returned from the \code{prep_data()} function.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1100 rows and 8 columns.
}
\usage{
test_chat_data
}
\description{
Just some random content inspired by a german rapper.
}
\keyword{datasets}
